requirements.txt
appdirs==1.4.3
Django==3.2.5
psycopg2-binary==2.8.6
django-crispy-forms==1.11.0
Pillow==8.3.1
django-filter==2.4.0
